add deploy

+++ Jenkinsfile
@@ -10,46 +10,40 @@ pipeline {
     stages {
 
         stage('Setup Tools') {
-            steps {
-                sh '''
-                    # kubectl
-                    curl -LO "https://dl.k8s.io/release/v1.29.0/bin/linux/amd64/kubectl"
-                    chmod +x kubectl
-                    cp kubectl /tmp/kubectl
+                    steps {
+                        sh '''
+                            # kubectl
+                            curl -LO "https://dl.k8s.io/release/v1.29.0/bin/linux/amd64/kubectl"
+                            chmod +x kubectl
+                            cp kubectl /tmp/kubectl
 
-                    # Docker CLI
-                    curl -fsSL https://download.docker.com/linux/static/stable/x86_64/docker-27.3.1.tgz -o /tmp/d.tgz
-                    cd /tmp && tar xzf d.tgz
-                    chmod +x docker/docker
-                    cp docker/docker /tmp/docker
-                '''
-            }
-        }
+                            # Docker CLI
+                            if [ ! -f /tmp/docker/docker ]; then
+                                curl -fsSL https://download.docker.com/linux/static/stable/x86_64/docker-27.3.1.tgz -o /tmp/d.tgz
+                                cd /tmp && tar xzf d.tgz
+                                chmod +x docker/docker
+                            fi
+                        '''
+                    }
+                }
 
-        stage('Checkout WalletApp') {
-            steps {
-                git branch: 'master',
-                    url: 'https://github.com/IvanJavDev/WalletApp.git'
-            }
-        }
+                stage('Maven Build') {
+                    steps {
+                        sh '''
+                            tar czf /tmp/project.tar.gz .
+                            cat /tmp/project.tar.gz | /tmp/docker/docker -H ${DOCKER_HOST} run --rm -i \
+                              -w /app \
+                              maven:3.9.9-eclipse-temurin-17-alpine \
+                              sh -c "cd /app && tar xzf - && mvn clean package -DskipTests"
+                        '''
+                    }
+                }
 
-        stage('Maven Build') {
-            steps {
-                sh '''
-                    tar czf /tmp/project.tar.gz .
-                    cat /tmp/project.tar.gz | /tmp/docker -H ${DOCKER_HOST} run --rm -i \
-                      -w /app \
-                      maven:3.9.9-eclipse-temurin-17-alpine \
-                      sh -c "cd /app && tar xzf - && mvn clean package -DskipTests"
-                '''
-            }
-        }
-
-        stage('Docker Build') {
-            steps {
-                sh '/tmp/docker -H ${DOCKER_HOST} build -t ${IMAGE_NAME} .'
-            }
-        }
+                stage('Docker Build') {
+                    steps {
+                        sh '/tmp/docker/docker -H ${DOCKER_HOST} build -t ${IMAGE_NAME} .'
+                    }
+                }
 
         stage('Deploy WalletApp') {
             steps {
